+++ docs/sidebars.js
@@ -22,6 +22,7 @@ module.exports = {
           'plugins_api/ImagePlugin',
           'plugins_api/DividerPlugin',
           'plugins_api/FileUploadPlugin',
+          'plugins_api/TextColorAndHighlightPlugins',
         ],
       },
       {