+++ docs/sidebars.js
@@ -17,6 +17,7 @@ module.exports = {
           'plugins_api/GiphyPlugin',
           'plugins_api/HashtagPlugin',
           'plugins_api/VideoPlugin',
+          'plugins_api/CodeBlockPlugin',
           'plugins_api/MapPlugin',
           'plugins_api/GalleryPlugin',
           'plugins_api/ImagePlugin',