+++ packages/plugin-vertical-embed/web/src/modals.js
@@ -1,11 +1,11 @@
-import PostSelectionInputModal from './toolbar/postSelectionInputModal';
+import VerticalEmbedInputModal from './toolbar/VerticalEmbedInputModal';
 
 const Modals = {
   POST_SELECTION_INPUT: 'post-selection-input',
 };
 
 const ModalsMap = {
-  [Modals.POST_SELECTION_INPUT]: PostSelectionInputModal,
+  [Modals.POST_SELECTION_INPUT]: VerticalEmbedInputModal,
 };
 
 export { Modals, ModalsMap };