+++ packages/plugin-vertical-embed/web/src/toolbar/VerticalEmbedInputModal.jsx
@@ -1,14 +1,20 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
-import { UrlInputModal, FOOTER_BUTTON_ALIGNMENT } from 'wix-rich-content-plugin-commons';
+import {
+  UrlInputModal,
+  FOOTER_BUTTON_ALIGNMENT,
+  LoaderIcon,
+} from 'wix-rich-content-plugin-commons';
 import { contentTypeMap } from '../constants';
 import ItemsList from './itemsList/ItemsList';
 import styles from '../../statics/styles/vertical-embed-modal.scss';
-export default class PostSelectionInputModal extends Component {
+import generalStyles from '../../statics/styles/general.scss';
+export default class VerticalEmbedInputModal extends Component {
   state = {
     errorMsg: '',
     products: [],
-    selectedProduct: null,
+    selectedProduct: this.props.componentData?.selectedProduct || null,
+    status: 'LOADING',
   };
 
   componentDidMount() {
@@ -18,11 +24,15 @@ export default class PostSelectionInputModal extends Component {
       locale,
     } = this.props;
     this.verticalApi = verticalsApi(type, locale);
-    this.verticalApi.search('').then(products => this.setState({ products }));
+    this.verticalApi.search('').then(products => {
+      this.setState({ products, status: products.length === 0 ? 'NO_ITEMS' : 'READY' });
+    });
   }
 
   onInputChange = (inputString = '') => {
-    this.verticalApi.search(inputString).then(products => this.setState({ products }));
+    this.verticalApi.search(inputString).then(products => {
+      this.setState({ products, status: products.length === 0 ? 'NOT_FOUND' : 'READY' });
+    });
     this.setState({ inputString });
   };
 
@@ -50,7 +60,7 @@ export default class PostSelectionInputModal extends Component {
   };
 
   render() {
-    const { products, inputString, selectedProduct } = this.state;
+    const { products, inputString, selectedProduct, status } = this.state;
     const {
       t,
       componentData: { type },
@@ -58,6 +68,20 @@ export default class PostSelectionInputModal extends Component {
       isMobile,
     } = this.props;
     const contentType = contentTypeMap[type];
+    const selected = selectedProduct !== null;
+    const emptyState = (
+      <div className={generalStyles.emptyState}>
+        <div className={generalStyles.title}>
+          {t(`Embed_Vertical_${contentType}_EmptyState_NoResults_Title`)}
+        </div>
+        <div className={generalStyles.description}>
+          {t(`Embed_Vertical_${contentType}_EmptyState_NoResults_Description`)}
+        </div>
+      </div>
+    );
+    const show = status !== 'NO_ITEMS';
+    const textInput = show ? { searchIcon: true } : false;
+
     return (
       <UrlInputModal
         onConfirm={this.onConfirm}
@@ -71,20 +95,32 @@ export default class PostSelectionInputModal extends Component {
         input={inputString}
         isMobile={isMobile}
         buttonAlignment={FOOTER_BUTTON_ALIGNMENT.END}
+        selected={selected}
+        textInput={textInput}
       >
         <div className={styles.itemsWrapper}>
-          <ItemsList
-            selectedItem={selectedProduct}
-            products={products}
-            onClick={this.onItemClick}
-          />
+          {status === 'LOADING' ? (
+            <div className={generalStyles.emptyState}>
+              <LoaderIcon className={styles.fileLoaderIcon} />
+            </div>
+          ) : status === 'NOT_FOUND' ? (
+            emptyState
+          ) : (
+            <ItemsList
+              selectedItem={selectedProduct}
+              products={products}
+              onClick={this.onItemClick}
+              contentType={contentType}
+              t={t}
+            />
+          )}
         </div>
       </UrlInputModal>
     );
   }
 }
 
-PostSelectionInputModal.propTypes = {
+VerticalEmbedInputModal.propTypes = {
   onConfirm: PropTypes.func,
   onReplace: PropTypes.func,
   helpers: PropTypes.object.isRequired,