+++ packages/plugin-vertical-embed/web/src/toolbar/insert-buttons.tsx
@@ -5,7 +5,7 @@ import {
   getModalStyles,
 } from 'wix-rich-content-editor-common';
 import { EventIcon, ProductIcon, BookingIcon } from '../icons';
-import PostSelectionInputModal from './postSelectionInputModal';
+import VerticalEmbedInputModal from './VerticalEmbedInputModal';
 import { DEFAULTS, contentTypeMap } from '../constants';
 import getModalCustomStyles from './ModalCustomStyles';
 import { CreateInsertButtons } from 'wix-rich-content-common';
@@ -33,7 +33,7 @@ const createInsertButtons: CreateInsertButtons<'t' | 'settings' | 'isMobile' | '
       Icon: iconsMap[type],
       componentData: { ...DEFAULTS, type },
       section: 'BlockToolbar_Section_Embed_Wix',
-      modalElement: decorateComponentWithProps(PostSelectionInputModal, { ...settings, locale }),
+      modalElement: decorateComponentWithProps(VerticalEmbedInputModal, { ...settings, locale }),
       modalStyles: getModalStyles({
         customStyles: getModalCustomStyles(isMobile),
         fullScreen: false,