+++ packages/viewer/web/src/RichContentViewer.jsx
@@ -126,7 +126,6 @@ class RichContentViewer extends Component {
       const { styles } = this;
       const {
         textDirection,
-        direction,
         decorators,
         inlineStyleMappers,
         locale,
@@ -163,7 +162,7 @@ class RichContentViewer extends Component {
       );
       return (
         <GlobalContext.Provider value={{ isMobile, t }}>
-          <div className={wrapperClassName} dir={direction || getLangDir(locale)}>
+          <div className={wrapperClassName} dir={getLangDir(locale)}>
             <div className={editorClassName}>{output}</div>
             <AccessibilityListener isMobile={this.props.isMobile} />
           </div>
@@ -203,7 +202,6 @@ RichContentViewer.propTypes = {
   relValue: PropTypes.string,
   config: PropTypes.object,
   textDirection: PropTypes.oneOf(['rtl', 'ltr']),
-  direction: PropTypes.oneOf(['rtl', 'ltr']),
   textAlignment: PropTypes.oneOf(['left', 'right']),
   disabled: PropTypes.bool,
   seoMode: PropTypes.oneOfType([PropTypes.bool, PropTypes.object]),