+++ packages/viewer/web/src/utils/convertContentState.jsx
@@ -86,14 +86,7 @@ const getBlocks = (mergedStyles, textDirection, context, addAnchorsPrefix) => {
           ? getInteractionWrapper({ interactions, context })
           : DefaultInteractionWrapper;
 
-        const textClassName = getBlockStyleClasses(
-          mergedStyles,
-          textDirection || blockProps.data[i]?.textDirection,
-          alignment
-        );
-
         const _child = isEmptyBlock(child) ? <br /> : child;
-
         const inner = (
           <ChildTag
             id={`viewer-${blockProps.keys[i]}`}
@@ -113,9 +106,7 @@ const getBlocks = (mergedStyles, textDirection, context, addAnchorsPrefix) => {
             style={blockDataToStyle(blockProps.data[i])}
             key={blockProps.keys[i]}
           >
-            <span style={{ display: 'block' }} className={textClassName}>
-              {_child}
-            </span>
+            {_child}
           </ChildTag>
         );