+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 ### :book: Documentation
 - `text-color`
   - [#1654](https://github.com/wix-incubator/rich-content/pull/1654) Text Color and Highlight Plugins Docs
+- `code-block`
+  - [#1640](https://github.com/wix-incubator/rich-content/pull/1640) Code Block Plugin Docs
 
 </details>
 <hr/>